+++ Framework/Built_In_Automation/Sequential_Actions/action_declarations/info.py
@@ -75,6 +75,7 @@ action_support = (
     "attribute constrain",
     "optional option",
     "graphql",
+    "shared capability"
 )
 
 # List of supported mobile platforms - must be lower case