+++ Framework/MainDriverApi.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # -*- coding: cp1252 -*-
-
+import copy
 import json
 import inspect
 import os
@@ -29,6 +29,14 @@ from Framework.Built_In_Automation.Shared_Resources import (
 )
 from Framework.Utilities import ws
 from reporting import junit_report
+
+from rich.style import Style
+from rich.table import Table
+from rich.console import Console
+from rich.box import ASCII_DOUBLE_HEAD, DOUBLE
+from rich.padding import Padding
+
+rich_print = Console().print
 
 top_path = os.path.dirname(os.getcwd())
 drivers_path = os.path.join(top_path, "Drivers")
@@ -171,7 +179,7 @@ def check_if_other_machines_failed_in_linked_run():
 
 
 # downloads attachments for a test case
-def create_tc_log_ss_folder(run_id, test_case, temp_ini_file):
+def create_tc_log_ss_folder(run_id, test_case, temp_ini_file, server_version):
     try:
         log_file_path = ConfigModule.get_config_value(
             "sectionOne", "temp_run_file_path", temp_ini_file
@@ -181,11 +189,11 @@ def create_tc_log_ss_folder(run_id, test_case, temp_ini_file):
     test_case_folder = (
         log_file_path +
         os.sep +
-        (run_id.replace(":", "-") +
+        run_id.replace(":", "-") +
         os.sep +
         CommonUtil.current_session_name +
         os.sep +
-        test_case.replace(":", "-"))
+        test_case.replace(":", "-")
     )
     # create test_case_folder
     ConfigModule.add_config_value("sectionOne", "test_case", test_case, temp_ini_file)
@@ -206,6 +214,12 @@ def create_tc_log_ss_folder(run_id, test_case, temp_ini_file):
     performance_report = test_case_folder + os.sep + "performance_report"
     ConfigModule.add_config_value("sectionOne", "performance_report", performance_report, temp_ini_file)
     FL.CreateFolder(performance_report)
+
+    if float(server_version.split(".")[0]) >= 7:
+        # json report folder
+        json_report = test_case_folder + os.sep + "json_report"
+        ConfigModule.add_config_value("sectionOne", "json_report", json_report, temp_ini_file)
+        FL.CreateFolder(json_report)
 
     # create where attachments from selenium browser will be downloaded
     zeuz_download_folder = test_case_folder + os.sep + "zeuz_download_folder"
@@ -502,6 +516,7 @@ def run_all_test_steps_in_a_test_case(
             current_step_name = all_step_info[StepSeq - 1]["step_name"]
             current_step_id = all_step_info[StepSeq - 1]["step_id"]
             current_step_sequence = all_step_info[StepSeq - 1]["step_sequence"]
+            shared.Set_Shared_Variables("zeuz_current_step", all_step_info[StepSeq - 1], print_variable=False, pretty=False)
 
             step_attachments = all_step_info[StepSeq - 1]['step_attachments']
             step_attachment_list = []
@@ -520,10 +535,19 @@ def run_all_test_steps_in_a_test_case(
             )
             CommonUtil.current_step_no = str(current_step_sequence)
             # add log
-            log_line = "STEP #%d: %s" % (StepSeq, current_step_name)
-            print("-"*len(log_line))
-            CommonUtil.ExecLog(sModuleInfo, log_line, 4)
-            print("-"*len(log_line))
+            # log_line = "STEP #%d: %s" % (StepSeq, current_step_name)
+            # print("-"*len(log_line))
+            # CommonUtil.ExecLog(sModuleInfo, log_line, 4)
+            # print("-"*len(log_line))
+
+            _color = "yellow"
+            # _style = Style(color="yellow", blink=False, bold=True)
+            table = Table(border_style=_color, box=ASCII_DOUBLE_HEAD, expand=False)
+            table.add_column(f"STEP-{str(StepSeq)}", justify="center", style=_color, max_width=40)
+            table.add_row(f"{current_step_name}", style=_color)
+            # width_pad = CommonUtil.max_char // 2 - (max(len(current_step_name), 6) + 4) // 2
+            # table = Padding(table, (0, width_pad))
+            rich_print(table)
 
             test_steps_data = all_step_dataset[StepSeq-1]
             test_action_info = all_action_info[StepSeq-1]
@@ -543,19 +567,24 @@ def run_all_test_steps_in_a_test_case(
             sTestStepStartTime = datetime.fromtimestamp(TestStepStartTime).strftime("%Y-%m-%d %H:%M:%S.%f")
             WinMemBegin = CommonUtil.PhysicalAvailableMemory()  # get available memory
 
-            sStepResult = call_driver_function_of_test_step(
-                sModuleInfo,
-                all_step_info,
-                StepSeq,
-                step_time,
-                current_step_name,
-                final_dependency,
-                final_run_params,
-                test_steps_data,
-                test_action_info,
-                file_specific_steps,
-                debug_actions,
-            )
+            if StepSeq in CommonUtil.disabled_step:
+                CommonUtil.ExecLog(sModuleInfo, "STEP-%s is disabled" % StepSeq, 2)
+                sStepResult = "skipped"
+            else:
+                sStepResult = call_driver_function_of_test_step(
+                    sModuleInfo,
+                    all_step_info,
+                    StepSeq,
+                    step_time,
+                    current_step_name,
+                    final_dependency,
+                    final_run_params,
+                    test_steps_data,
+                    test_action_info,
+                    file_specific_steps,
+                    debug_actions,
+                )
+
             TestStepEndTime = time.time()
             sTestStepEndTime = datetime.fromtimestamp(TestStepEndTime).strftime("%Y-%m-%d %H:%M:%S.%f")
             WinMemEnd = CommonUtil.PhysicalAvailableMemory()  # get available memory
@@ -592,8 +621,12 @@ def run_all_test_steps_in_a_test_case(
                 CommonUtil.ExecLog(sModuleInfo, "%s : Test Step Passed" % current_step_name, 1)
                 after_execution_dict.update({"status": PASSED_TAG})
 
+            elif sStepResult.upper() == "SKIPPED":
+                CommonUtil.ExecLog(sModuleInfo, "%s : Test Step Skipped" % current_step_name, 2)
+                after_execution_dict.update({"status": "Skipped"})
+
             else:
-                CommonUtil.ExecLog(sModuleInfo, "%s%s" % (current_step_name, CommonUtil.to_dlt_from_fail_reason), 3)  # add log
+                CommonUtil.ExecLog(sModuleInfo, "%s%s" % (current_step_name, CommonUtil.to_dlt_from_fail_reason), 3)
                 after_execution_dict.update({"status": "Failed"})  # dictionary update
 
                 # check if set for continue
@@ -672,7 +705,7 @@ def calculate_test_case_result(sModuleInfo, TestCaseID, run_id, sTestStepResultL
 
 
 # writes the log file for a test case
-def zip_and_delete_tc_folder(
+def zip_and_delete_tc_folder_old(
     sTestCaseStatus,
     temp_ini_file,
     send_log_file_only_for_fail=True,
@@ -683,6 +716,33 @@ def zip_and_delete_tc_folder(
             FL.ZipFolder(
                 ConfigModule.get_config_value("sectionOne", "test_case_folder", temp_ini_file),
                 ConfigModule.get_config_value("sectionOne", "test_case_folder", temp_ini_file) + ".zip",
+            )
+    # Delete the folder
+    FL.DeleteFolder(ConfigModule.get_config_value("sectionOne", "test_case_folder", temp_ini_file))
+
+
+# writes the log file for a test case
+def zip_and_delete_tc_folder(
+    run_id,
+    TestCaseID,
+    sTestCaseStatus,
+    temp_ini_file,
+    send_log_file_only_for_fail=True,
+
+):
+    # if settings checked, then send log file or screenshots, otherwise don't send
+    if sTestCaseStatus not in passed_tag_list or sTestCaseStatus in passed_tag_list and not send_log_file_only_for_fail:
+        if ConfigModule.get_config_value("RunDefinition", "local_run") == "False":
+            all_steps = CommonUtil.all_logs_json[CommonUtil.runid_index]["test_cases"][CommonUtil.tc_index]["steps"]
+            for step in all_steps:
+                json_filename = Path(ConfigModule.get_config_value("sectionOne", "json_report", temp_ini_file))/(str(step["step_sequence"])+".json")
+                with open(json_filename, "w") as f:
+                    json.dump(step, f)
+
+            zip_name = run_id.replace(":", "-") + "_" + TestCaseID.replace(":", "-") + ".zip"
+            FL.ZipFolder(
+                ConfigModule.get_config_value("sectionOne", "test_case_folder", temp_ini_file),
+                str(Path(ConfigModule.get_config_value("sectionOne", "test_case_folder", temp_ini_file)).parent/zip_name),
             )
     # Delete the folder
     FL.DeleteFolder(ConfigModule.get_config_value("sectionOne", "test_case_folder", temp_ini_file))
@@ -729,12 +789,11 @@ def run_test_case(
     final_dependency,
     final_run_params,
     temp_ini_file,
-    is_linked,
     testcase_info,
-    executor,
     debug_info,
     all_file_specific_steps,
     rerun_on_fail,
+    server_version,
     send_log_file_only_for_fail=True,
     performance=False,
     browserDriver=None,
@@ -746,14 +805,24 @@ def run_test_case(
         CommonUtil.current_tc_no = test_case
         CommonUtil.load_testing = False
         ConfigModule.add_config_value("sectionOne", "sTestStepExecLogId", sModuleInfo, temp_ini_file)
-        create_tc_log_ss_folder(run_id, test_case, temp_ini_file)
+        create_tc_log_ss_folder(run_id, test_case, temp_ini_file, server_version)
         set_important_variables()
         file_specific_steps = all_file_specific_steps[TestCaseID] if TestCaseID in all_file_specific_steps else {}
         TestCaseName = testcase_info["title"]
-        log_line = "# EXECUTING TEST CASE : %s :: %s #" % (test_case, TestCaseName)
-        print("#"*(len(log_line)))
-        CommonUtil.ExecLog("", log_line, 4, False)
-        print("#"*(len(log_line)))
+        shared.Set_Shared_Variables("zeuz_current_tc", testcase_info, print_variable=False, pretty=False)
+
+        # log_line = "# EXECUTING TEST CASE : %s :: %s #" % (test_case, TestCaseName)
+        # print("#"*(len(log_line)))
+        # CommonUtil.ExecLog("", log_line, 4, False)
+        # print("#"*(len(log_line)))
+        _color = "white"
+        # danger_style = Style(color=_color, blink=False, bold=True)
+        table = Table(border_style=_color, box=DOUBLE, expand=False, padding=1)
+        table.add_column(test_case, justify="center", style=_color, max_width=40)
+        table.add_row(TestCaseName, style=_color)
+        # width_pad = CommonUtil.max_char//2 - (max(len(TestCaseName), len(test_case)) + 4)//2
+        # table = Padding(table, (0, width_pad))
+        rich_print(table)
 
         # get test case start time
         if performance and browserDriver:
@@ -829,15 +898,25 @@ def run_test_case(
 
         if not CommonUtil.debug_status:  # if normal run, then write log file and cleanup driver instances
             CommonUtil.Join_Thread_and_Return_Result("screenshot")  # Let the capturing screenshot end in thread
-            cleanup_driver_instances()  # clean up drivers
+            if shared.Get_Shared_Variables("zeuz_auto_teardown").strip().lower() in ("on", "yes", "true", "ok", "enable"):
+                cleanup_driver_instances()  # clean up drivers
             shared.Clean_Up_Shared_Variables()  # clean up shared variables
             if ConfigModule.get_config_value("RunDefinition", "local_run") == "False":
-                zip_and_delete_tc_folder(
-                    sTestCaseStatus,
-                    temp_ini_file,
-                    send_log_file_only_for_fail
-                )
 
+                if float(server_version.split(".")[0]) < 7:
+                    zip_and_delete_tc_folder_old(
+                        sTestCaseStatus,
+                        temp_ini_file,
+                        send_log_file_only_for_fail
+                    )
+                else:
+                    zip_and_delete_tc_folder(
+                        run_id,
+                        TestCaseID,
+                        sTestCaseStatus,
+                        temp_ini_file,
+                        send_log_file_only_for_fail
+                    )
         return "passed"
     except:
         CommonUtil.Exception_Handler(sys.exc_info())
@@ -1011,7 +1090,7 @@ def check_run_cancel(run_id):
     # CommonUtil.run_cancelled = False
 
 
-def upload_json_report(Userid, temp_ini_file, run_id):
+def upload_json_report_old(Userid, temp_ini_file, run_id):
     try:
         if CommonUtil.debug_status: return
         zip_path = Path(ConfigModule.get_config_value("sectionOne", "temp_run_file_path", temp_ini_file))/run_id.replace(":", "-")/CommonUtil.current_session_name
@@ -1080,6 +1159,107 @@ def upload_json_report(Userid, temp_ini_file, run_id):
         CommonUtil.Exception_Handler(sys.exc_info())
 
 
+def upload_reports_and_zips(Userid, temp_ini_file, run_id):
+    try:
+        if CommonUtil.debug_status: return
+        zip_dir = Path(ConfigModule.get_config_value("sectionOne", "temp_run_file_path", temp_ini_file))/run_id.replace(":", "-")/CommonUtil.current_session_name
+
+        if ConfigModule.get_config_value("RunDefinition", "local_run") == "False" and CommonUtil.run_cancel != CANCELLED_TAG:
+            # FL.ZipFolder(str(zip_path), str(zip_path) + ".zip")
+
+            tc_report = copy.deepcopy(CommonUtil.all_logs_json)
+            tc_report[CommonUtil.runid_index]["machine_name"] = Userid
+
+            for testcase in tc_report[CommonUtil.runid_index]["test_cases"]:
+                for step in testcase["steps"]:
+                    if "actions" in step:
+                        del step["actions"]
+                    if "log" in step:
+                        del step["log"]
+
+            with open("report.json", "w") as f:
+                f.write(json.dumps({"execution_report": json.dumps(tc_report)}))
+
+            for _ in range(5):
+                try:
+                    res = requests.post(
+                        RequestFormatter.form_uri("create_report_log_api/"),
+                        data={"execution_report": json.dumps(tc_report)},
+                        verify=False,
+                        **RequestFormatter.add_api_key_to_headers({}))
+                    if res.status_code == 200:
+                        print(f"Successfully uploaded the execution report of run_id {run_id}")
+                        break
+                    else:
+                        print(f"Failed to upload the execution report of run_id {run_id}")
+                        print(f"Status: {res.status_code} Response: {res.text}")
+                        print("Retrying...")
+                    time.sleep(4)
+                except:
+                    CommonUtil.Exception_Handler(sys.exc_info())
+                    time.sleep(4)
+            else:
+                print("Could not Upload the report to server of run_id '%s'" % run_id)
+
+            zip_files = [os.path.join(zip_dir, f) for f in os.listdir(zip_dir) if f.endswith(".zip")]
+            opened_zips = []
+            size = 0
+            for zip_file in zip_files:
+                opened_zips.append(open(str(zip_file), "rb"))
+                size += round(os.stat(str(zip_file)).st_size / 1024, 2)
+
+            if size > 1024:
+                size = str(round(size/1024, 3)) + " MB"
+            else:
+                size = str(round(size, 3)) + " KB"
+            print("Uploading %s logs-screenshots of %s testcases of %s from:\n%s" % (CommonUtil.current_session_name, len(zip_files), size, str(zip_dir)))
+
+            for _ in range(5):
+                try:
+                    files_list = []
+                    for zips in opened_zips:
+                        files_list.append(("file",zips))
+                    res = requests.post(
+                        RequestFormatter.form_uri("save_log_and_attachment_api/"),
+                        files=files_list,
+                        data={"machine_name": Userid},
+                        verify=False,
+                        **RequestFormatter.add_api_key_to_headers({}))
+                    if res.status_code == 200:
+                        try:
+                            res_json = res.json()
+                        except:
+                            # print("Could not Upload logs-screenshots to server")
+                            # print("\nResponse Text = " + res.text + "\n")
+                            # break
+                            continue
+                        if isinstance(res_json, dict) and 'message' in res_json and res_json["message"]:
+                            print("Successfully Uploaded logs-screenshots to server of run_id '%s'" % run_id)
+                        else:
+                            print("Could not Upload logs-screenshots to server of run_id '%s'" % run_id)
+                            print("\nResponse Text = " + res.text + "\n")
+                        break
+                except:
+                    pass
+
+                time.sleep(4)
+            else:
+                print("Could not Upload logs-screenshots to server of run_id '%s'" % run_id)
+
+        with open(zip_dir / "execution_log_old_format.json", "w") as f:
+            json.dump(CommonUtil.get_all_logs(json=True), f, indent=2)
+
+        if CommonUtil.run_cancel != CANCELLED_TAG:
+            # Create a standard report format to be consumed by other tools.
+            junit_report_path = zip_dir / "junitreport.xml"
+            print("Generating junit4 compatible report.")
+            junit_report.process(CommonUtil.all_logs_json, str(junit_report_path))
+            print("DONE. Generated junit report at %s" % junit_report_path)
+        return zip_dir
+    except:
+        CommonUtil.Exception_Handler(sys.exc_info())
+
+
 def split_testcases(run_id_info, max_tc_in_single_session):
     import copy
     from math import ceil, floor
@@ -1125,7 +1305,7 @@ def main(device_dict, user_info_object):
         Userid = (CommonUtil.MachineInfo().getLocalUser()).lower()
 
         get_json, all_file_specific_steps = True, {}
-        save_path = Path(temp_ini_file).parent / "attachments"
+        save_path = Path(ConfigModule.get_config_value("sectionOne", "temp_run_file_path", temp_ini_file)) / "attachments"
         cnt = 0
         for i in os.walk(save_path):
             if get_json:
@@ -1151,12 +1331,12 @@ def main(device_dict, user_info_object):
         for run_id_info in all_run_id_info:
             run_id_info["base_path"] = ConfigModule.get_config_value("Advanced Options", "_file_upload_path")
             run_id = run_id_info["run_id"]
-            release_version = run_id_info["release_version"]
+            server_version = run_id_info["release_version"]
             release_name = run_id_info["release_name"]
             release_info = run_id_info["release_info"]
             CommonUtil.ExecLog(
                 "",
-                "Server version = %s\nServer Release Date = %s\nServer Release Note = %s" % (release_version, release_name, release_info),
+                "Server version = %s\nServer Release Date = %s\nServer Release Note = %s" % (server_version, release_name, release_info),
                 4,
                 False,
             )
@@ -1173,23 +1353,23 @@ def main(device_dict, user_info_object):
 
             # Start websocket server if we're in debug mode.
             if run_id.lower().startswith("debug"):
-                CommonUtil.ExecLog(
-                    "",
-                    "\n********************************\n*    STARTING DEBUG SESSION    *\n********************************",
-                    4,
-                    False,
-                )
+                # CommonUtil.ExecLog(
+                #     "",
+                #     "\n********************************\n*    STARTING DEBUG SESSION    *\n********************************",
+                #     4,
+                #     False,
+                # )
                 CommonUtil.debug_status = True
                 print("[LIVE LOG] Connecting to Live Log service")
                 ws.connect()
                 print("[LIVE LOG] Connected to Live Log service")
             else:
-                CommonUtil.ExecLog(
-                    "",
-                    "\n******************************\n*    STARTING RUN SESSION    *\n******************************",
-                    4,
-                    False,
-                )
+                # CommonUtil.ExecLog(
+                #     "",
+                #     "\n******************************\n*    STARTING RUN SESSION    *\n******************************",
+                #     4,
+                #     False,
+                # )
                 CommonUtil.debug_status = False
                 cleanup_driver_instances()  # clean up drivers
                 shared.Clean_Up_Shared_Variables()  # clean up shared variables
@@ -1240,6 +1420,7 @@ def main(device_dict, user_info_object):
             rerun_on_fail = ConfigModule.get_config_value("RunDefinition", "rerun_on_fail")
             rerun_on_fail = False if rerun_on_fail.lower() == "false" else True
             CommonUtil.upload_on_fail, CommonUtil.rerun_on_fail = send_log_file_only_for_fail, rerun_on_fail
+            shared.Set_Shared_Variables("zeuz_auto_teardown", "on")
 
             all_testcases_info = run_id_info["test_cases"]
             TestSetStartTime = time.time()
@@ -1260,7 +1441,7 @@ def main(device_dict, user_info_object):
             num_of_tc = len(all_testcases_info)
             cnt = 1
 
-            max_tc_in_single_session = 25    # Todo: make it 25
+            max_tc_in_single_session = 50    # Todo: make it 25
             all_sessions = split_testcases(run_id_info, max_tc_in_single_session)
             session_cnt = 1
             for each_session in all_sessions:
@@ -1276,6 +1457,7 @@ def main(device_dict, user_info_object):
                     test_case_no = testcase_info["testcase_no"]
                     test_case_name = testcase_info["title"]
                     set_device_info_according_to_user_order(device_order, device_dict, test_case_no, test_case_name, user_info_object, Userid)
+                    CommonUtil.disabled_step = []
 
                     if performance_test_case:
                         # get performance test info
@@ -1363,12 +1545,11 @@ def main(device_dict, user_info_object):
                             final_dependency,
                             final_run_params,
                             temp_ini_file,
-                            is_linked,
                             testcase_info,
-                            executor,
                             debug_info,
                             all_file_specific_steps,
                             rerun_on_fail,
+                            server_version,
                             send_log_file_only_for_fail,
                         )
                         CommonUtil.clear_all_logs()  # clear logs
@@ -1392,7 +1573,12 @@ def main(device_dict, user_info_object):
                     "duration": TestSetDuration
                 }
                 CommonUtil.CreateJsonReport(setInfo=after_execution_dict)
-                upload_json_report(Userid, temp_ini_file, run_id)
+
+                if float(server_version.split(".")[0]) < 7:
+                    upload_json_report_old(Userid, temp_ini_file, run_id)
+                else:
+                    upload_reports_and_zips(Userid, temp_ini_file, run_id)
+
                 session_cnt += 1
 
             print("Test set execution time = %s sec for %s testcases" % (round(TimeDiff, 3), num_of_tc))